break up config_parser private tests
